--- examples/grpc_example/Package.resolved
@@ -1,14 +1,0 @@
-{
-  "pins" : [
-    {
-      "identity" : "swift-log",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/apple/swift-log",
-      "state" : {
-        "revision" : "532d8b529501fb73a2455b179e0bbb6d49b652ed",
-        "version" : "1.5.3"
-      }
-    }
-  ],
-  "version" : 2
-}
