+++ examples/grpc_example/Package.swift
@@ -5,7 +5,6 @@ import PackageDescription
 let package = Package(
     name: "grpc_example",
     dependencies: [
-        // TODO: Replace this dependency with your dependencies.
-        .package(url: "https://github.com/apple/swift-log", from: "1.5.2"),
+        // .package(url: "https://github.com/grpc/grpc-swift", from: "1.16.0"),
     ]
 )
