--- examples/grpc_example/main.swift
@@ -1,4 +1,0 @@
-import Logging
-
-let logger = Logger(label: "com.example.main")
-logger.info("Hello World!")
