+++ examples/grpc_example/sources/server/server_main.swift
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import Dispatch
+import Foundation
 import GRPC
 import NIOCore
 import NIOPosix
