+++ examples/grpc_example/swift_deps_index.json
@@ -1,38 +1,5 @@
 {
-  "direct_dep_identities": [
-    "swift-log"
-  ],
-  "modules": [
-    {
-      "name": "Logging",
-      "c99name": "Logging",
-      "src_type": "swift",
-      "label": "@swiftpkg_swift_log//:Sources_Logging",
-      "package_identity": "swift-log",
-      "product_memberships": [
-        "Logging"
-      ]
-    }
-  ],
-  "products": [
-    {
-      "identity": "swift-log",
-      "name": "Logging",
-      "type": "library",
-      "target_labels": [
-        "@swiftpkg_swift_log//:Sources_Logging"
-      ]
-    }
-  ],
-  "packages": [
-    {
-      "name": "swiftpkg_swift_log",
-      "identity": "swift-log",
-      "remote": {
-        "commit": "532d8b529501fb73a2455b179e0bbb6d49b652ed",
-        "remote": "https://github.com/apple/swift-log",
-        "version": "1.5.3"
-      }
-    }
-  ]
+  "direct_dep_identities": [],
+  "modules": [],
+  "products": []
 }
