+++ gazelle/BUILD.bazel
@@ -19,7 +19,6 @@ go_library(
     name = "gazelle",
     srcs = [
         "config.go",
-        "constants.go",
         "doc.go",
         "generate.go",
         "kinds.go",
@@ -41,7 +40,6 @@ go_library(
         "@bazel_gazelle//config:go_default_library",
         "@bazel_gazelle//label:go_default_library",
         "@bazel_gazelle//language:go_default_library",
-        "@bazel_gazelle//language/proto:go_default_library",
         "@bazel_gazelle//repo:go_default_library",
         "@bazel_gazelle//resolve:go_default_library",
         "@bazel_gazelle//rule:go_default_library",
