--- gazelle/constants.go
@@ -1,11 +1,0 @@
-package gazelle
-
-const (
-	// swiftProtoSuffix is the suffix applied to the labels of all generated
-	// swift_proto_library targets.
-	swiftProtoSuffix = "_swift_proto"
-
-	// swiftGRPCSuffix is the suffix applied to the labels of all generated
-	// swift_grpc_library targets.
-	swiftGRPCSuffix = "_swift_grpc"
-)
