+++ gazelle/internal/swift/BUILD.bazel
@@ -35,6 +35,7 @@ go_library(
         "rule_attr_string_dict.go",
         "rule_kinds.go",
         "rules_common.go",
+        "rules_from_protos.go",
         "rules_from_srcs.go",
         "test_rules_common.go",
     ],
@@ -47,6 +48,7 @@ go_library(
         "@bazel_gazelle//config:go_default_library",
         "@bazel_gazelle//label:go_default_library",
         "@bazel_gazelle//language:go_default_library",
+        "@bazel_gazelle//language/proto:go_default_library",
         "@bazel_gazelle//rule:go_default_library",
         "@com_github_bazelbuild_buildtools//build:go_default_library",
         "@com_github_deckarep_golang_set_v2//:go_default_library",
